requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
streamlit-authenticator
pyyaml
l